Use address instead of slot as context handle for ToNode funcs
For pipeline commands the knowledge that a command has been sent is
stored using the command struct. This struct contains the slot where
the command was sent to, and is stored in a list used when attempting
to get the responses.
When using CLUSTER NODES for route updates its not certain that every master
node serves slots. When using CLUSTER SLOTS it should be for certain.

Lets use the address, a string containing the host:port, instead which can
be used to lookup the node and its hiredis-context.

The ToNode() async case will now use an own callback that will skip
redirects/retries, so slot is not needed here either.